requirements.txt
python-telegram-bot[job-queue]==20.7
flask
motor==3.6.0
pymongo==4.9.2
python-dotenv==1.0.1
dnspython==2.6.1
aiohttp==3.10.10
async-timeout==4.0.3